feat: DiscountCalculator 의 역할을 CouponRepository 가 하도록 변경
