+++ app/src/main/java/woowacourse/shopping/ui/payment/PaymentViewModel.kt
@@ -11,7 +11,6 @@ import woowacourse.shopping.common.SingleLiveData
 import woowacourse.shopping.common.UniversalViewModelFactory
 import woowacourse.shopping.data.coupon.remote.CouponRemoteRepository
 import woowacourse.shopping.data.order.remote.OrderRemoteRepository
-import woowacourse.shopping.domain.model.coupon.DiscountCalculator
 import woowacourse.shopping.domain.repository.coupon.CouponRepository
 import woowacourse.shopping.domain.repository.order.OrderRepository
 import woowacourse.shopping.ui.ResponseHandler.handleResponseResult
@@ -24,8 +23,6 @@ class PaymentViewModel(
     private val orderRepository: OrderRepository,
     private val couponRepository: CouponRepository,
 ): ViewModel(), OnCouponClickListener {
-    private var discountCalculator: DiscountCalculator = DiscountCalculator(orderInformation, emptyList())
-
     private val _couponsUiModel = MutableLiveData<List<CouponUiModel>>(emptyList())
     val couponsUiModel: LiveData<List<CouponUiModel>> get() = _couponsUiModel
 
@@ -56,10 +53,8 @@ class PaymentViewModel(
 
     fun loadCoupons() {
         viewModelScope.launch {
-            handleResponseResult(couponRepository.loadCoupons(), _errorMessage) { coupons ->
-                discountCalculator = DiscountCalculator(orderInformation, coupons)
-                val applicableCoupon = discountCalculator.findAvailableCoupons(orderInformation.cartItems)
-                _couponsUiModel.value = applicableCoupon.map { CouponMapper.toUiModel(it) }
+            handleResponseResult(couponRepository.loadCoupons(orderInformation), _errorMessage) { coupons ->
+                _couponsUiModel.value = coupons.map { CouponMapper.toUiModel(it) }
             }
         }
     }
@@ -78,9 +73,9 @@ class PaymentViewModel(
         }
 
         val isChecked = !couponsUiModel.first { it.id == couponId }.checked
-        _discountAmount.value = discountCalculator.calculateDiscountAmount(couponId, isChecked)
-        _shippingFee.value = discountCalculator.calculateShippingFee(couponId, isChecked)
-        _totalPaymentAmount.value = discountCalculator.calculateTotalAmount(couponId, isChecked)
+        _discountAmount.value = couponRepository.calculateDiscountAmount(orderInformation, couponId, isChecked)
+        _shippingFee.value = couponRepository.calculateShippingFee(orderInformation, couponId, isChecked)
+        _totalPaymentAmount.value = couponRepository.calculateTotalAmount(orderInformation, couponId, isChecked)
     }
 
     companion object {
